% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baydec-package.R
\docType{package}
\name{baydec-package}
\alias{baydec-package}
\alias{baydec}
\title{The 'baydec' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
