% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{deconvole}
\usage{
deconvolve(X, Y, tol = 0.05, iter = 1000, cores = 4, chains = 4, ...)
}
\arguments{
\item{X}{the signature: a genes x cells matrix}

\item{Y}{the signal: a genes x replicates matrix}

\item{tol}{the sd in the likelihood}

\item{iter}{number of iterations in the chains}

\item{cores}{number of cores used (e.g. = no. cores)}

\item{chains}{number of MCMC chains}

\item{...}{}
}
\value{
a stan sample object
}
\description{
MCMC sampling for deconvolution
}
\details{
This generates a stan sample from the model
}
\author{
Pete Dodd
}
