% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{postprocess.baydec}
\alias{postprocess.baydec}
\title{Post-process}
\usage{
postprocess.baydec(X)
}
\arguments{
\item{X}{stan sample object}
}
\value{
a summary of samples
}
\description{
Extract proportions from samples
}
\details{
This generates summary statistics from MCMC samples.
}
\author{
Pete Dodd
}
