% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baydec-package.R
\docType{package}
\name{baydec-package}
\alias{baydec}
\alias{baydec-package}
\title{baydec: Bayesian Deconvolution}
\description{
Bayesian Deconvolution using MCMC, with cytometry data in mind.
}
\author{
\strong{Maintainer}: Pete Dodd \email{p.j.dodd@sheffield.ac.uk} (\href{https://orcid.org/0000-0001-5825-9347}{ORCID})

}
\keyword{internal}
